cleaning up duck types
