+++ lib/minesweeper_2pl.rb
@@ -4,7 +4,8 @@ require "minesweeper_2pl/cli"
 require "minesweeper_2pl/game"
 require "minesweeper_2pl/board"
 require "minesweeper_2pl/board_formatter"
-require "minesweeper_2pl/cat_emoji_formatter"
+require "minesweeper_2pl/cat_emoji"
+require "minesweeper_2pl/bomb_emoji"
 require "minesweeper_2pl/board_printer"
 require "minesweeper_2pl/player"
 require "minesweeper_2pl/messages"
