+++ lib/minesweeper_2pl/bomb_emoji.rb
@@ -1,0 +1,15 @@
+module Minesweeper
+  class BombEmoji
+    def show_lost_emoji
+      "\u{1f4a3}"
+    end
+
+    def show_won_emoji
+      "\u{1f3c6}"
+    end
+
+    def show_flag_emoji
+      "\u{1f6a9}"
+    end
+  end
+end
