+++ lib/minesweeper_2pl/cat_emoji.rb
@@ -1,0 +1,15 @@
+module Minesweeper
+  class CatEmoji
+    def show_lost_emoji
+      "\u{1f640}"
+    end
+
+    def show_won_emoji
+      "\u{1f63a}"
+    end
+
+    def show_flag_emoji
+      "\u{1f364}"
+    end
+  end
+end
